+++ core/src/com/fruit/tests/Dummy.java
@@ -65,9 +65,9 @@ public class Dummy extends com.fruit.logic.objects.abstracted.Character {
     }
 
     @Override
-    public void changeHealthPoints(int amount){
+    public void changeHealthPoints(float amount){
         super.changeHealthPoints(amount);
-        Controller.addOnScreenMessage(Integer.toString(amount), getBody().getPosition().x * PIXELS_TO_METERS,
+        Controller.addOnScreenMessage(Float.toString(amount), getBody().getPosition().x * PIXELS_TO_METERS,
                 getBody().getPosition().y * PIXELS_TO_METERS, 1.5f);
     }
 }
